Rip remaining c-api portions into gif-c
